Refactors project for dependency injection
