+++ package.json
@@ -33,6 +33,8 @@
     "sequelize-cli": "^6.3.0"
   },
   "devDependencies": {
-    "nodemon": "^2.0.14"
+    "jest": "^27.3.1",
+    "nodemon": "^2.0.14",
+    "supertest": "^6.1.6"
   }
 }
